Use lib.rs and Make checkbox
